backend/requirements.txt
# Django and REST Framework
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1
django-oauth-toolkit>=2.3.0
django-cors-headers>=4.3.1
django-csp==3.7

# Channels and WebSocket
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0
websockets>=12.0

# HTTP and Async
aiohttp>=3.9.0
requests>=2.31.0

# Database and Caching
redis>=5.0.1

# Static Files
whitenoise>=6.6.0

# Environment and Configuration
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-django>=4.8.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Code Quality
black>=23.7.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.5.1
